Use more memory
This branch is meant for internal use at Go About only. It may be publically visible, but is not meant for public consumption. If you see this commit show up elsewhere, somebody messed up. This branch may be rewound, deleted or other wise messed up at any time and for any reason.